05_assistants_code_interpreter/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["khalidch73 <khalidnawazch73@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
requests = "^2.31.0"
aiokafka = "^0.10.0"
protobuf = "^5.27.1"
python-dotenv = "^1.0.1"
openai = "^1.33.0"





[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
